Changing size_t to int and updating some checks

+++ connectfour.h
@@ -8,13 +8,13 @@
 #include <stdbool.h>
 struct matrix
 {
-    size_t rows, cols;
+    int rows, cols;
     int* data;
 };
 extern void PrintGrid(struct matrix* m);
-extern bool IsMoveValid(struct matrix* m, size_t c);
+extern bool IsMoveValid(struct matrix* m, int c);
 extern bool Win(struct matrix* m, int value);
-extern size_t GetPlayerMove(struct matrix*m);
-extern void Place(struct matrix* m, int value, size_t c);
+extern int GetPlayerMove(struct matrix*m);
+extern void Place(struct matrix* m, int value, int c);
 extern void ConnectFour(struct matrix* m);
 #endif //CONNECTFOUR2_CONNECTFOUR_H
